User commit

+++ src/main/java/com/example/demo/entity/Student.java
@@ -1,20 +1,20 @@
 package com.example.demo.entity;
 
 public class Student {
-    private int id;
+    private long int id;
     private String name;
     private String email;
     
-    public Student(int id, String name, String email) {
+    public Student(long int id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
     }
 
-    public int getId() {
+    public long int getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(long int id) {
         this.id = id;
     }
     public String getName() {
